GoogleSearch_WS/requirements.txt
google-generativeai
python-dotenv
requests
beautifulsoup4
playwright[browsers]
pandas
openpyxl